requirements.txt
boto3==1.38.38
botocore==1.38.38
dnspython==2.7.0
et_xmlfile==2.0.0
jmespath==1.0.1
numpy==2.3.0
openpyxl==3.1.5
pandas==2.3.0
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
s3transfer==0.13.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
